refactor(workspace-runtime): extract runtime event owner
