+++ src/renderer/src/lib/acp/useWorkspaceAgentRuntime.ts
@@ -1,7 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 
 import type {
-  AcpConnectionStatus,
   AcpCreateSessionResponse,
   AcpPermissionGrant,
   AcpPermissionRequest,
@@ -46,10 +45,14 @@ import {
   type HistoryReplayDescriptor
 } from './history-preamble'
 import {
-  applyWorkspaceRuntimeEvent,
-  syncWorkspaceAgentFirstOutputState,
+  createWorkspaceRuntimeEventProcessor,
+  drainWorkspaceRuntimeEventsForPersistence,
+  markRunningSessionsDisconnectedOnDrop,
+  processVisibleWorkspaceRuntimeEvents,
+  processWorkspaceRuntimeEvents,
+  syncWorkspaceContextUsage,
   syncWorkspacePermissionState
-} from './workspace-events'
+} from './workspace-runtime-event-owner'
 
 type SendWorkspaceMessageInput = {
   sessionId?: string
@@ -186,13 +189,6 @@ type WorkspacePermissionProfileRuntime = Pick<
   'state' | 'setPermissionProfile'
 >
 type PersistSessionDeletion = (request: { projectId: string; sessionId: string }) => Promise<void>
-
-type RuntimeEventApplier = (event: AcpRuntimeEvent) => Promise<boolean>
-
-type WorkspaceRuntimeEventProcessor = {
-  process: (events: AcpRuntimeEvent[]) => Promise<void>
-  drain: (sessionId?: string) => Promise<void>
-}
 
 // Runtime adoption and Branch reset intentionally complete before appendUserMessage creates the next
 // run. Keep duplicate preparations closed during that idle-looking interval without exposing a
@@ -450,197 +446,6 @@ const shutdownNotebookForBranchChange = async (
   if (typeof window === 'undefined' || !window.api?.notebook?.shutdown) return
   await window.api.notebook.shutdown({ sessionId, workspaceCwd, projectName })
 }
-
-const processVisibleWorkspaceRuntimeEvents = async (
-  events: AcpRuntimeEvent[],
-  processedEventIds: Set<string>,
-  applyEvent: RuntimeEventApplier = applyWorkspaceRuntimeEvent,
-  processingEventIds = new Set<string>()
-): Promise<void> => {
-  // Runtime snapshots are bounded, so forget ids that can no longer be replayed from the source list.
-  const visibleEventIds = new Set(events.map((event) => event.id))
-
-  for (const eventId of processedEventIds) {
-    if (!visibleEventIds.has(eventId)) {
-      processedEventIds.delete(eventId)
-    }
-  }
-
-  for (const eventId of processingEventIds) {
-    if (!visibleEventIds.has(eventId)) {
-      processingEventIds.delete(eventId)
-    }
-  }
-
-  for (const event of events) {
-    if (processedEventIds.has(event.id) || processingEventIds.has(event.id)) continue
-
-    processingEventIds.add(event.id)
-    try {
-      // Apply visible events sequentially so message chunks and artifact finalization stay ordered.
-      await applyEvent(event)
-      processedEventIds.add(event.id)
-    } catch {
-      // Artifact finalization errors are recorded by the adapter before throwing.
-      // Keeping this id unprocessed lets the same visible runtime event retry.
-      continue
-    } finally {
-      processingEventIds.delete(event.id)
-    }
-  }
-}
-
-const createWorkspaceRuntimeEventProcessor = (
-  applyEvent: RuntimeEventApplier = applyWorkspaceRuntimeEvent
-): WorkspaceRuntimeEventProcessor => {
-  type EventLane = {
-    acceptedEvents: Map<string, AcpRuntimeEvent>
-    failedEventIds: Set<string>
-    processedEventIds: Set<string>
-    processingEventIds: Set<string>
-    drainInFlight?: Promise<void>
-    drainAgain: boolean
-  }
-
-  const unscopedEventLane = Symbol('unscoped-workspace-runtime-events')
-  const eventLanes = new Map<string | symbol, EventLane>()
-  let latestEvents: AcpRuntimeEvent[] = []
-  let acceptedEventVersion = 0
-
-  const getEventLaneKey = (event: AcpRuntimeEvent): string | symbol =>
-    event.sessionId ?? unscopedEventLane
-
-  const getEventLane = (laneKey: string | symbol): EventLane => {
-    let lane = eventLanes.get(laneKey)
-    if (!lane) {
-      lane = {
-        acceptedEvents: new Map<string, AcpRuntimeEvent>(),
-        failedEventIds: new Set<string>(),
-        processedEventIds: new Set<string>(),
-        processingEventIds: new Set<string>(),
-        drainAgain: false
-      }
-      eventLanes.set(laneKey, lane)
-    }
-
-    return lane
-  }
-
-  const cleanEventLane = (laneKey: string | symbol, lane: EventLane): void => {
-    const visibleEventIds = new Set(
-      latestEvents.filter((event) => getEventLaneKey(event) === laneKey).map((event) => event.id)
-    )
-
-    for (const eventId of lane.acceptedEvents.keys()) {
-      if (!visibleEventIds.has(eventId) && lane.processedEventIds.has(eventId)) {
-        lane.acceptedEvents.delete(eventId)
-        lane.failedEventIds.delete(eventId)
-        lane.processedEventIds.delete(eventId)
-        lane.processingEventIds.delete(eventId)
-      }
-    }
-
-    if (lane.acceptedEvents.size === 0 && !lane.drainInFlight) {
-      eventLanes.delete(laneKey)
-    }
-  }
-
-  const drainLane = async (laneKey: string | symbol): Promise<void> => {
-    const lane = getEventLane(laneKey)
-
-    if (lane.drainInFlight) {
-      lane.drainAgain = true
-      return lane.drainInFlight
-    }
-
-    lane.drainInFlight = (async () => {
-      do {
-        lane.drainAgain = false
-        await processVisibleWorkspaceRuntimeEvents(
-          [...lane.acceptedEvents.values()],
-          lane.processedEventIds,
-          async (event) => {
-            const hadFailed = lane.failedEventIds.has(event.id)
-            try {
-              const applied = await applyEvent(event)
-              lane.failedEventIds.delete(event.id)
-              return applied
-            } catch (error) {
-              const isVisible = latestEvents.some(
-                (candidate) => candidate.id === event.id && getEventLaneKey(candidate) === laneKey
-              )
-              if (hadFailed && !isVisible) {
-                lane.acceptedEvents.delete(event.id)
-                lane.failedEventIds.delete(event.id)
-              } else {
-                lane.failedEventIds.add(event.id)
-              }
-              throw error
-            }
-          },
-          lane.processingEventIds
-        )
-      } while (lane.drainAgain)
-    })()
-
-    try {
-      await lane.drainInFlight
-    } finally {
-      lane.drainInFlight = undefined
-      cleanEventLane(laneKey, lane)
-    }
-  }
-
-  return {
-    process: (events) => {
-      latestEvents = events
-      const visibleLaneKeys = new Set<string | symbol>()
-
-      for (const event of events) {
-        const laneKey = getEventLaneKey(event)
-        const lane = getEventLane(laneKey)
-        visibleLaneKeys.add(laneKey)
-
-        if (
-          !lane.processedEventIds.has(event.id) &&
-          !lane.processingEventIds.has(event.id) &&
-          !lane.acceptedEvents.has(event.id)
-        ) {
-          // A bounded source snapshot may evict this event before a slow predecessor finishes.
-          lane.acceptedEvents.set(event.id, event)
-          acceptedEventVersion += 1
-        }
-      }
-
-      for (const [laneKey, lane] of eventLanes) {
-        cleanEventLane(laneKey, lane)
-      }
-
-      const drains = [...visibleLaneKeys].map((laneKey) => drainLane(laneKey))
-      for (const [laneKey, lane] of eventLanes) {
-        if (!visibleLaneKeys.has(laneKey) && lane.acceptedEvents.size > 0) {
-          void drainLane(laneKey)
-        }
-      }
-
-      return Promise.all(drains).then(() => undefined)
-    },
-    drain: async (sessionId) => {
-      if (sessionId !== undefined) {
-        if (eventLanes.has(sessionId)) await drainLane(sessionId)
-        return
-      }
-
-      let drainedVersion: number
-      do {
-        drainedVersion = acceptedEventVersion
-        await Promise.all([...eventLanes.keys()].map((laneKey) => drainLane(laneKey)))
-      } while (drainedVersion !== acceptedEventVersion)
-    }
-  }
-}
-
-const liveWorkspaceRuntimeEventProcessor = createWorkspaceRuntimeEventProcessor()
 
 const sessionOwnsActivePrompt = (sessionId: string, messageId: string): boolean => {
   const session = useSessionStore
@@ -1920,64 +1725,6 @@ const processContextOverflowRecovery = (
   }
 }
 
-// Flags sessions with an active prompt or native compaction as disconnected on a TRANSITION into a
-// dropped connection state. Abnormal drops (agent crash / gateway drop) go through main's
-// handleConnectionClosed → status 'closed'; deliberate mid-prompt disconnects use disconnect(false)
-// (no 'closed' emit) and idle provider/skills reconnects have no active interaction, so neither reaches
-// markDisconnected here.
-const markRunningSessionsDisconnectedOnDrop = (
-  previousStatus: AcpConnectionStatus,
-  currentStatus: AcpConnectionStatus,
-  previousSessionStatuses: Partial<Record<string, AcpConnectionStatus>> = {},
-  currentSessionStatuses: Partial<Record<string, AcpConnectionStatus>> = {}
-): void => {
-  const { sessions, markDisconnected } = useSessionStore.getState()
-
-  for (const session of sessions) {
-    if (
-      session.status !== 'running' &&
-      session.status !== 'waiting-permission' &&
-      !session.compacting
-    ) {
-      continue
-    }
-
-    const previousOwnedStatus = previousSessionStatuses[session.id]
-    const currentOwnedStatus = currentSessionStatuses[session.id]
-    const hasOwningRuntimeStatus =
-      previousOwnedStatus !== undefined || currentOwnedStatus !== undefined
-    const previous = hasOwningRuntimeStatus
-      ? (previousOwnedStatus ?? currentOwnedStatus ?? previousStatus)
-      : previousStatus
-    const current = hasOwningRuntimeStatus
-      ? (currentOwnedStatus ?? previousOwnedStatus ?? currentStatus)
-      : currentStatus
-    const droppedNow =
-      (current === 'closed' || current === 'error') && previous !== 'closed' && previous !== 'error'
-
-    if (droppedNow) markDisconnected(session.id)
-  }
-}
-
-// Copies live context usage into the durable Session. Missing usage clears only attached sessions,
-// preserving the last snapshot for detached sessions while invalidating replaced runtime contexts.
-const syncWorkspaceContextUsage = (
-  sessionIds: readonly string[],
-  contextUsageBySession: Record<string, AcpContextUsage>
-): void => {
-  const { setContextUsage } = useSessionStore.getState()
-  for (const sessionId of sessionIds) {
-    setContextUsage(sessionId, contextUsageBySession[sessionId])
-  }
-}
-
-const drainWorkspaceRuntimeEventsForPersistence = async (sessionId?: string): Promise<void> => {
-  const snapshot = await window.api.acp.getState()
-  void liveWorkspaceRuntimeEventProcessor.process(snapshot.events)
-  await liveWorkspaceRuntimeEventProcessor.drain(sessionId)
-  syncWorkspaceContextUsage(snapshot.sessionIds, snapshot.contextUsageBySession)
-}
-
 // Deletes in three ordered ownership layers: agent runtime, durable JSON/DB coordinator, then renderer
 // state. A failure in either authoritative layer leaves the session visible with an actionable error.
 const deleteWorkspaceSession = async (
@@ -2133,8 +1880,7 @@ const useWorkspaceAgentRuntime = (): {
   // Publish ownership before processing the same snapshot's events. A first visible chunk then clears
   // the wait monotonically instead of a later effect rearming it from that snapshot.
   useEffect(() => {
-    syncWorkspaceAgentFirstOutputState(agentPromptInFlightSessionIds)
-    void liveWorkspaceRuntimeEventProcessor.process(runtime.state.events)
+    void processWorkspaceRuntimeEvents(runtime.state.events, agentPromptInFlightSessionIds)
   }, [agentPromptInFlightSessionIds, runtime.state.events])
 
   // Mirrors pending permission requests into per-session store status.
